Create model factories and add setup main seeder

+++ database/migrations/2023_05_10_103333_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->decimal('price', 8, 2);
-            $table->float('availability')->default(0);
+            $table->integer('availability')->default(0);
             $table->timestamps();
         });
     }
