+++ database/migrations/2023_05_10_103405_create_order_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('order_id');
             $table->unsignedBigInteger('product_id');
-            $table->float('quantity')->default(0);
+            $table->integer('quantity')->default(0);
             $table->decimal('price', 8, 2)->default(0);
             $table->timestamps();
